Add fairseq trained model translation option
This is a first step to incorporating custom machine learning tl.
Based on MingShiba's trained model for VNOCR.

From my testing it does better at some things but is overall worse.
A few tls in test files have been updated but they are hand picked.

However, the possibility is there now for any model to be used.