Add color coding to mdb's skill data
Also don't write size in the extract script, defer to text processing.

I always thought about doing this but now that TLG requires newlines
in the text files, it's not possible to use the previous newline
separation for multi-effect skills (like Kitasan's) and without
color it becomes confusing to read.
This way saves space anyway so that's another positive.